refactor: move shared utilities, hooks, and types from plugin-core to ui-core
Per principal dev feedback, plugin-core should only contain plugin/extension
framework code. Shared utilities, hooks, and type helpers belong in ui-core.
