+++ frontend/src/app/featureFlags/useDevFeatureFlags.ts
@@ -4,7 +4,7 @@ import { useSearchParams } from 'react-router-dom';
 import { useExtensions } from '@odh-dashboard/plugin-core';
 import { isAreaExtension } from '@odh-dashboard/plugin-core/extension-points';
 import type { DashboardCommonConfig, DashboardConfigKind } from '@odh-dashboard/k8s-core';
-import { useBrowserStorage } from '@odh-dashboard/plugin-core/utilities';
+import { useBrowserStorage } from '@odh-dashboard/ui-core/utilities';
 import { definedFeatureFlags, SupportedAreasStateMap } from '#~/concepts/areas/const';
 import { DevFeatureFlags } from '#~/types';
 import axios from '#~/utilities/axios';
