+++ frontend/src/pages/modelServing/screens/metrics/bias/useBiasChartSelections.ts
@@ -1,6 +1,6 @@
 import { useParams } from 'react-router-dom';
 import React from 'react';
-import { useBrowserStorage } from '@odh-dashboard/plugin-core/utilities';
+import { useBrowserStorage } from '@odh-dashboard/ui-core/utilities';
 import { BiasMetricConfig } from '#~/concepts/trustyai/types';
 import { byId } from '#~/pages/modelServing/screens/metrics/utils';
