+++ frontend/src/pages/projects/notebook/useStopNotebookModalAvailability.ts
@@ -1,4 +1,4 @@
-import { useBrowserStorage } from '@odh-dashboard/plugin-core/utilities';
+import { useBrowserStorage } from '@odh-dashboard/ui-core/utilities';
 
 const useStopNotebookModalAvailability = (): [boolean, (v: boolean) => void] =>
   useBrowserStorage<boolean>('odh.dashboard.dsg.stop.modal.preference', false);
