+++ frontend/src/pages/projects/screens/detail/storage/useStorageStatusAlert.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import type { PersistentVolumeClaimKind } from '@odh-dashboard/k8s-core';
-import { useBrowserStorage } from '@odh-dashboard/plugin-core/utilities';
+import { useBrowserStorage } from '@odh-dashboard/ui-core/utilities';
 import useNotification from '#~/utilities/useNotification';
 import { getDisplayNameFromK8sResource } from '#~/concepts/k8s/utils';
 import { getFullStatusFromPercentage } from './utils';
