+++ packages/model-registry/src/modelCatalog/ValidatedModelsBanner.tsx
@@ -10,7 +10,7 @@ import {
   Title,
 } from '@patternfly/react-core';
 import { TimesIcon } from '@patternfly/react-icons';
-import { useBrowserStorage } from '@odh-dashboard/plugin-core/utilities';
+import { useBrowserStorage } from '@odh-dashboard/ui-core/utilities';
 import validatedModelsBannerImg from '@odh-dashboard/internal/images/validated-models-banner.svg';
 
 const BANNER_STORAGE_KEY = 'odh.dashboard.modelCatalog.validatedModelsBanner.dismissed';
