+++ packages/model-serving/src/components/deploymentWizard/dynamicFormUtils.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useResolvedExtensions } from '@odh-dashboard/plugin-core';
-import type { RecursivePartial } from '@odh-dashboard/plugin-core/utilities';
+import type { RecursivePartial } from '@odh-dashboard/ui-core/utilities';
 import type { DeploymentWizardFieldOverride, WizardField, WizardFormData } from './types';
 import {
   isDeploymentWizardFieldOverrideExtension,
