+++ packages/model-serving/src/components/deploymentWizard/fields/NumReplicasField.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { FormGroup, HelperText, HelperTextItem } from '@patternfly/react-core';
 import { z } from 'zod';
 import NumberInputWrapper from '@odh-dashboard/internal/components/NumberInputWrapper';
-import { normalizeBetween } from '@odh-dashboard/plugin-core/utilities';
+import { normalizeBetween } from '@odh-dashboard/ui-core/utilities';
 
 // Schema
 const LOWER_LIMIT = 1;
