+++ packages/model-serving/src/components/deploymentWizard/fields/modelLocationFields/URIConnectionField.tsx
@@ -12,7 +12,7 @@ import {
   UriField,
 } from '@odh-dashboard/internal/concepts/connectionTypes/types';
 import DefaultValueTextRenderer from '@odh-dashboard/internal/concepts/connectionTypes/fields/DefaultValueTextRenderer';
-import { trimInputOnBlur, trimInputOnPaste } from '@odh-dashboard/plugin-core/utilities';
+import { trimInputOnBlur, trimInputOnPaste } from '@odh-dashboard/ui-core/utilities';
 import { FieldMode } from '@odh-dashboard/internal/concepts/connectionTypes/fields/types';
 
 const validateUrl = (url?: string) => {
