+++ packages/model-serving/src/concepts/useStopModalPreference.ts
@@ -1,4 +1,4 @@
-import { useBrowserStorage } from '@odh-dashboard/plugin-core/utilities';
+import { useBrowserStorage } from '@odh-dashboard/ui-core/utilities';
 
 export const STOP_MODAL_PREFERENCE_KEY = 'odh.dashboard.modelServing.stop.modal.preference';
 const useStopModalPreference = (): [boolean, (v: boolean) => void] =>
