+++ packages/model-training/src/global/trainingJobList/hooks/usePauseTrainingJobModalAvailability.ts
@@ -1,4 +1,4 @@
-import { useBrowserStorage } from '@odh-dashboard/plugin-core/utilities';
+import { useBrowserStorage } from '@odh-dashboard/ui-core/utilities';
 
 const usePauseTrainingJobModalAvailability = (): [boolean, (v: boolean) => void] =>
   useBrowserStorage<boolean>('odh.dashboard.model-training.pause.modal.preference', false);
