+++ packages/plugin-core/package.json
@@ -15,16 +15,7 @@
     "./helpers/ui": "./src/core/helpers/ui.ts",
     "./extension-points": "./src/extension-points/index.ts",
     "./areas": "./src/areas/index.ts",
-    "./testing": "./src/testing/utils.ts",
-    "./utilities": "./src/utilities/index.ts",
-    "./utilities/const": "./src/utilities/const.ts",
-    "./utilities/string": "./src/utilities/string.ts",
-    "./utilities/trimInput": "./src/utilities/trimInput.ts",
-    "./utilities/valueUnits": "./src/utilities/valueUnits.ts",
-    "./utilities/utils": "./src/utilities/utils.ts",
-    "./typeHelpers": "./src/types/typeHelpers.ts",
-    "./hooks": "./src/hooks/index.ts",
-    "./hooks/useBrowserStorage": "./src/hooks/useBrowserStorage.tsx"
+    "./testing": "./src/testing/utils.ts"
   },
   "dependencies": {
     "@odh-dashboard/k8s-core": "*",
@@ -35,8 +26,7 @@
     "@odh-dashboard/eslint-config": "*",
     "@odh-dashboard/jest-config": "*",
     "@odh-dashboard/tsconfig": "*",
-    "@types/lodash-es": "^4.17.12",
-    "@types/node": "^22.0.0"
+    "@types/lodash-es": "^4.17.12"
   },
   "peerDependencies": {
     "@patternfly/react-core": "^6",
