+++ packages/ui-core/src/utilities/string.ts
