+++ packages/ui-core/src/utilities/trimInput.ts
